Add a MIT license as discussed with the author.
I have emailed Gareth and requested that the sources be released under an
open source license like MIT, and they have agreed. The new LICENSE file
has been added here as required.

The other references will be cleaned up in the source code so that we
can build things as we need to in the future. I really grateful to
Gareth for releasing the source so that it can be updated for modern
systems and built into cross-compilation tools.