Update LICENSE to be explicitly MIT license.
Gareth had said that it should be exactly the usual MIT license text, and
whilst the preceding text was as supplied by him, it wasn't recognised as
MIT  because of the clarity he'd added for the 'mkdrawf software'. As he
said it's acceptable to use the usual MIT license, I've used the MIT
license text as given by GitHub.

The difference is purely in the text of 'the software' and the word
wrapping.