Fix for incorrect references to VersionNum.
I'd put VersionNum at the top of the repository, which made sense, but
it seems that I couldn't get it to link properly for the dependencies
in amu - I am certain I've got the filename translations broken in there
somehow. Anyhow, moving the VersionNum to alongside the source works
better anyhow.