Tentative implementation for RISC OS 64.
May now build for RISC OS 64. Added new functions for the gstrans
and other calls that were in the syscalls assembler file. This should
allow us to handle the operations in a relatively system way if we're
running on RISC OS 64, and we may even run on non-RISC OS 64 with the
manual implementation of GSTrans.